backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
python-dotenv
pydantic
python-multipart
pillow
pillow-heif
passlib[bcrypt]
bcrypt==4.1.2
python-jose[cryptography]
pytest
httpx
asyncpg
email-validator
apscheduler
